+++ packages/apollo-link-batch/src/batching.ts
@@ -2,7 +2,7 @@ import { Observable, Operation, NextLink, FetchResult } from 'apollo-link';
 
 export type BatchHandler = (
   operations: Operation[],
-  forward: (NextLink | undefined)[],
+  forward?: (NextLink | undefined)[],
 ) => Observable<FetchResult[]> | null;
 
 export interface BatchableRequest {
@@ -13,9 +13,9 @@ export interface BatchableRequest {
   // added to the queue and is resolved once the result is back
   // from the server.
   observable?: Observable<FetchResult>;
-  next?: (result: FetchResult) => void;
-  error?: (error: Error) => void;
-  complete?: () => void;
+  next?: Array<(result: FetchResult) => void>;
+  error?: Array<(error: Error) => void>;
+  complete?: Array<() => void>;
 }
 
 // QueryBatcher doesn't fire requests immediately. Requests that were enqueued within
@@ -23,64 +23,98 @@ export interface BatchableRequest {
 // into one query.
 export class OperationBatcher {
   // Queue on which the QueryBatcher will operate on a per-tick basis.
-  public queuedRequests: BatchableRequest[] = [];
+  // Public only for testing
+  public queuedRequests: Map<string, BatchableRequest[]>;
 
   private batchInterval: number;
   private batchMax: number;
 
   //This function is called to the queries in the queue to the server.
   private batchHandler: BatchHandler;
+  private batchKey: (operation: Operation) => string;
 
   constructor({
     batchInterval,
     batchMax = 0,
     batchHandler,
+    batchKey = () => '',
   }: {
     batchInterval: number;
     batchMax?: number;
     batchHandler: BatchHandler;
+    batchKey?: (operation: Operation) => string;
   }) {
-    this.queuedRequests = [];
+    this.queuedRequests = new Map();
     this.batchInterval = batchInterval;
     this.batchMax = batchMax;
     this.batchHandler = batchHandler;
+    this.batchKey = batchKey;
   }
 
   public enqueueRequest(request: BatchableRequest): Observable<FetchResult> {
-    const requestCopy = { ...request };
+    const requestCopy = {
+      ...request,
+    };
+    let queued = false;
 
-    requestCopy.observable =
-      requestCopy.observable ||
-      new Observable<FetchResult>(observer => {
-        this.queuedRequests.push(requestCopy);
+    const key = this.batchKey(request.operation);
 
-        requestCopy.next = requestCopy.next || observer.next.bind(observer);
-        requestCopy.error = requestCopy.error || observer.error.bind(observer);
-        requestCopy.complete =
-          requestCopy.complete || observer.complete.bind(observer);
+    if (!requestCopy.observable) {
+      requestCopy.observable = new Observable<FetchResult>(observer => {
+        if (!this.queuedRequests.has(key)) {
+          this.queuedRequests.set(key, []);
+        }
+
+        if (!queued) {
+          this.queuedRequests.get(key).push(requestCopy);
+          queued = true;
+        }
+
+        //called for each subscriber, so need to save all listeners(next, error, complete)
+        requestCopy.next = requestCopy.next || [];
+        if (observer.next) requestCopy.next.push(observer.next.bind(observer));
+
+        requestCopy.error = requestCopy.error || [];
+        if (observer.error)
+          requestCopy.error.push(observer.error.bind(observer));
+
+        requestCopy.complete = requestCopy.complete || [];
+        if (observer.complete)
+          requestCopy.complete.push(observer.complete.bind(observer));
 
         // The first enqueued request triggers the queue consumption after `batchInterval` milliseconds.
-        if (this.queuedRequests.length === 1) {
-          this.scheduleQueueConsumption();
+        if (this.queuedRequests.get(key).length === 1) {
+          this.scheduleQueueConsumption(key);
         }
 
         // When amount of requests reaches `batchMax`, trigger the queue consumption without waiting on the `batchInterval`.
-        if (this.queuedRequests.length === this.batchMax) {
-          this.consumeQueue();
+        if (this.queuedRequests.get(key).length === this.batchMax) {
+          this.consumeQueue(key);
         }
       });
+    }
 
     return requestCopy.observable;
   }
 
   // Consumes the queue.
   // Returns a list of promises (one for each query).
-  public consumeQueue(): (Observable<FetchResult> | undefined)[] | undefined {
-    const requests: Operation[] = this.queuedRequests.map(
+  public consumeQueue(
+    key: string = '',
+  ): (Observable<FetchResult> | undefined)[] | undefined {
+    const queuedRequests = this.queuedRequests.get(key);
+
+    if (!queuedRequests) {
+      return;
+    }
+
+    this.queuedRequests.delete(key);
+
+    const requests: Operation[] = queuedRequests.map(
       queuedRequest => queuedRequest.operation,
     );
 
-    const forwards: NextLink[] = this.queuedRequests.map(
+    const forwards: NextLink[] = queuedRequests.map(
       queuedRequest => queuedRequest.forward,
     );
 
@@ -88,37 +122,57 @@ export class OperationBatcher {
     const nexts: any[] = [];
     const errors: any[] = [];
     const completes: any[] = [];
-    this.queuedRequests.forEach((batchableRequest, index) => {
+    queuedRequests.forEach((batchableRequest, index) => {
       observables.push(batchableRequest.observable);
       nexts.push(batchableRequest.next);
       errors.push(batchableRequest.error);
       completes.push(batchableRequest.complete);
     });
 
-    this.queuedRequests = [];
-
     const batchedObservable =
       this.batchHandler(requests, forwards) || Observable.of();
 
+    const onError = error => {
+      //each callback list in batch
+      errors.forEach(rejecters => {
+        if (rejecters) {
+          //each subscriber to request
+          rejecters.forEach(e => e(error));
+        }
+      });
+    };
+
     batchedObservable.subscribe({
       next: results => {
+        if (!Array.isArray(results)) {
+          results = [results];
+        }
+
+        if (nexts.length !== results.length) {
+          const error = new Error(
+            `server returned results with length ${
+              results.length
+            }, expected length of ${nexts.length}`,
+          );
+          (error as any).result = results;
+
+          return onError(error);
+        }
+
         results.forEach((result, index) => {
+          // attach the raw response to the context for usage
+          requests[index].setContext({ response: result });
           if (nexts[index]) {
-            nexts[index](result);
+            nexts[index].forEach(next => next(result));
           }
         });
       },
-      error: error => {
-        errors.forEach((rejecter, index) => {
-          if (errors[index]) {
-            errors[index](error);
-          }
-        });
-      },
+      error: onError,
       complete: () => {
         completes.forEach(complete => {
           if (complete) {
-            complete();
+            //each subscriber to request
+            complete.forEach(c => c());
           }
         });
       },
@@ -127,10 +181,10 @@ export class OperationBatcher {
     return observables;
   }
 
-  private scheduleQueueConsumption(): void {
+  private scheduleQueueConsumption(key: string = ''): void {
     setTimeout(() => {
-      if (this.queuedRequests.length) {
-        this.consumeQueue();
+      if (this.queuedRequests.get(key) && this.queuedRequests.get(key).length) {
+        this.consumeQueue(key);
       }
     }, this.batchInterval);
   }