+++ packages/apollo-link-error/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ### vNEXT
 
+### 1.0.6
+- udate apollo link with zen-observable-ts [PR#515](https://github.com/apollographql/apollo-link/pull/515)
+
+### 1.0.5
+- ApolloLink upgrade
+
+### 1.0.4
+- ApolloLink upgrade
+
 ### 1.0.3
 - export options as named interface [TypeScript]